# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YouTube Music Service",
...
    "termsOfService": "https://example.com/terms/",
    "contact": {
      "name": "YouTube Music API Service",
      "url": "https://github.com/your-repo"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/browse/home": {
      "get": {
        "tags": [
          "browse",
          "browse"
        ],
        "summary": "Get home page",
        "description": "Obtiene el contenido de la página principal de YouTube Music.",
        "operationId": "get_home_api_v1_browse_home_get",
        "responses": {
          "200": {
            "description": "Contenido obtenido exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": true,
                  "error_code": "VALIDATION_ERROR",
                  "message": "Error de validación en 'video_id': ID de video inválido. Debe tener exactamente 11 caracteres.",
                  "details": {
                    "field": "video_id",
                    "reason": "invalid_length",
                    "expected_length": 11,
                    "actual_length": 5
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Error de autenticación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": true,
                  "error_code": "AUTHENTICATION_ERROR",
                  "message": "Error de autenticación con YouTube Music. Verifica las credenciales.",
                  "details": {
                    "operation": "search",
                    "hint": "Check server logs for details"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Recurso no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": true,
                  "error_code": "NOT_FOUND",
                  "message": "Video no encontrado.",
                  "details": {
                    "resource_type": "video",
                    "resource_id": "rMbATaj7Il8"
                  }
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests - Rate limit excedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": true,
                  "error_code": "RATE_LIMIT_ERROR",
                  "message": "Límite de peticiones excedido. Intenta más tarde.",
                  "details": {
                    "operation": "get_stream",
                    "retry_after": 300
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": true,
                  "error_code": "INTERNAL_ERROR",
                  "message": "Ha ocurrido un error interno. Por favor intenta más tarde."
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway - Error de servicio externo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": true,
                  "error_code": "EXTERNAL_SERVICE_ERROR",
                  "message": "Error en YouTube Music durante get_stream. Intenta más tarde.",
                  "details": {
                    "operation": "get_stream",
                    "service": "YouTube Music"
                  }
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable - Servicio no disponible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": true,
                  "error_code": "SERVICE_UNAVAILABLE",
                  "message": "Servicio temporalmente no disponible debido a sobrecarga.",
                  "details": {
                    "retry_after": 180,
                    "state": "OPEN"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/browse/artist/{channel_id}/albums": {
      "get": {
        "tags": [
          "browse",
          "browse"
        ],
        "summary": "Get artist albums",
        "description": "Obtiene todos los álbumes de un artista.",
        "operationId": "get_artist_albums_api_v1_browse_artist__channel_id__albums_get",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID del canal del artista",
              "examples": {
                "example1": {
                  "value": "UC..."
                }
              },
              "title": "Channel Id"
            },
            "description": "ID del canal del artista"
          },
          {
            "name": "params",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Parámetros de paginación",
              "title": "Params"
            },
            "description": "Parámetros de paginación"
          }
        ],
        "responses": {
          "200": {
            "description": "Álbumes obtenidos exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistAlbumsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Error de validación",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "VALIDATION_ERROR",
                  "message": "Error de validación en 'video_id': ID de video inválido. Debe tener exactamente 11 caracteres.",
                  "details": {
                    "field": "video_id",
                    "reason": "invalid_length",
                    "expected_length": 11,
                    "actual_length": 5
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Error de autenticación",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "AUTHENTICATION_ERROR",
                  "message": "Error de autenticación con YouTube Music. Verifica las credenciales.",
                  "details": {
                    "operation": "search",
                    "hint": "Check server logs for details"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Recurso no encontrado",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "NOT_FOUND",
                  "message": "Video no encontrado.",
                  "details": {
                    "resource_type": "video",
                    "resource_id": "rMbATaj7Il8"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests - Rate limit excedido",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "RATE_LIMIT_ERROR",
                  "message": "Límite de peticiones excedido. Intenta más tarde.",
                  "details": {
                    "operation": "get_stream",
                    "retry_after": 300
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "INTERNAL_ERROR",
                  "message": "Ha ocurrido un error interno. Por favor intenta más tarde."
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway - Error de servicio externo",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "EXTERNAL_SERVICE_ERROR",
                  "message": "Error en YouTube Music durante get_stream. Intenta más tarde.",
                  "details": {
                    "operation": "get_stream",
                    "service": "YouTube Music"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable - Servicio no disponible",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "SERVICE_UNAVAILABLE",
                  "message": "Servicio temporalmente no disponible debido a sobrecarga.",
                  "details": {
                    "retry_after": 180,
                    "state": "OPEN"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/browse/album/{album_id}": {
      "get": {
        "tags": [
          "browse",
          "browse"
        ],
        "summary": "Get album information",
        "description": "Obtiene información completa de un álbum incluyendo todas sus canciones.",
        "operationId": "get_album_api_v1_browse_album__album_id__get",
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID del álbum",
              "examples": {
                "example1": {
                  "value": "MPREb..."
                }
              },
              "title": "Album Id"
            },
            "description": "ID del álbum"
          },
          {
            "name": "include_stream_urls",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Incluir stream URLs y mejores thumbnails para tracks",
              "default": true,
              "title": "Include Stream Urls"
            },
            "description": "Incluir stream URLs y mejores thumbnails para tracks"
          }
        ],
        "responses": {
          "200": {
            "description": "Álbum obtenido exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumResponse"
                },
                "example": {
                  "title": "Album Title",
                  "artists": [
                    {
                      "name": "Artist"
                    }
                  ],
                  "tracks": [
                    {
                      "videoId": "rMbATaj7Il8",
                      "title": "Track Title",
                      "stream_url": "https://...",
                      "thumbnail": "https://..."
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Error de validación",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "VALIDATION_ERROR",
                  "message": "Error de validación en 'video_id': ID de video inválido. Debe tener exactamente 11 caracteres.",
                  "details": {
                    "field": "video_id",
                    "reason": "invalid_length",
                    "expected_length": 11,
                    "actual_length": 5
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Error de autenticación",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "AUTHENTICATION_ERROR",
                  "message": "Error de autenticación con YouTube Music. Verifica las credenciales.",
                  "details": {
                    "operation": "search",
                    "hint": "Check server logs for details"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Recurso no encontrado",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "NOT_FOUND",
                  "message": "Video no encontrado.",
                  "details": {
                    "resource_type": "video",
                    "resource_id": "rMbATaj7Il8"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests - Rate limit excedido",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "RATE_LIMIT_ERROR",
                  "message": "Límite de peticiones excedido. Intenta más tarde.",
                  "details": {
                    "operation": "get_stream",
                    "retry_after": 300
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "INTERNAL_ERROR",
                  "message": "Ha ocurrido un error interno. Por favor intenta más tarde."
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway - Error de servicio externo",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "EXTERNAL_SERVICE_ERROR",
                  "message": "Error en YouTube Music durante get_stream. Intenta más tarde.",
                  "details": {
                    "operation": "get_stream",
                    "service": "YouTube Music"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable - Servicio no disponible",
            "content": {
              "application/json": {
                "example": {
                  "error": true,
                  "error_code": "SERVICE_UNAVAILABLE",
                  "message": "Servicio temporalmente no disponible debido a sobrecarga.",
                  "details": {
                    "retry_after": 180,
                    "state": "OPEN"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtistAlbumsResponse": {
        "properties": {
          "results": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Results",
            "description": "List of albums"
          },
          "browse_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Browse Id",
            "description": "Browse ID"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ArtistAlbumsResponse",
        "description": "Response for artist albums endpoint."
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "boolean",
            "title": "Error",
            "description": "Always True for error responses",
            "default": true
          },
          "error_code": {
            "type": "string",
            "title": "Error Code",
            "description": "Machine-readable error code",
            "examples": [
              "VALIDATION_ERROR",
              "AUTHENTICATION_ERROR",
              "RATE_LIMIT_ERROR",
              "NOT_FOUND",
              "EXTERNAL_SERVICE_ERROR",
              "SERVICE_UNAVAILABLE",
              "INTERNAL_ERROR"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable error message in Spanish",
            "examples": [
              "Error de validación en los parámetros de la petición",
              "Error de autenticación con YouTube Music",
              "Límite de peticiones excedido. Intenta más tarde.",
              "Recurso no encontrado",
              "Error en el servicio de YouTube Music",
              "Servicio temporalmente no disponible"
            ]
          },
          "details": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetails"
              },
              {
                "type": "null"
              }
            ],
            "description": "Additional error details (structure varies by error type)"
          }
        },
        "type": "object",
        "required": [
          "error_code",
          "message"
        ],
        "title": "ErrorResponse",
        "description": "Standard error response schema for all API errors.\n\nThis schema is used consistently across all endpoints to provide\na predictable error format for clients.\n\nAttributes:\n    error: Always True for error responses.\n    error_code: Machine-readable error code for programmatic handling.\n    message: Human-readable error message in Spanish.\n    details: Optional additional details about the error."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HomeResponse": {
        "properties": {},
        "additionalProperties": true,
        "type": "object",
        "title": "HomeResponse",
        "description": "Response for home endpoint."
      },
...
    }
  }
}